mammo-scan-api/requirements.txt
pydicom==2.4.4
pyradiomics==3.1.0
numpy==1.26.4
pandas==2.3.2
opencv-python==4.11.0.86
scikit-learn==1.6.1
scikit-image==0.24.0
joblib==1.5.2
SimpleITK==2.5.2
PyYAML==6.0.2
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
python-gdcm